requirements.txt
boto3
python-ulid
pynamodb
pynamodb-attributes
